requirements.txt
Flask==2.2.2
pytest==7.2.0
coverage==6.5.0
Jinja2==3.1.2
MarkupSafe==2.1.1
itsdangerous==2.1.2
Werkzeug==2.2.2
click==8.1.3
numpy==1.23.4
requests==2.28.1
pathlib==1.0.1
pandas==1.4.4
surprise==0.1
pyyaml==6.0
kafka-python==2.0.2
scikit-learn==1.1.3 
sklearn==0.0 
threadpoolctl==3.1.0
pymongo==4.3.2
dnspython==2.2.1
gunicorn==20.1.0
gevent==22.10.2
wheel==0.38.3